refactor(platform): split HAL/PlatformProcess.h into Platform/<OS>/<OS>PlatformProcess.cpp (B4)
