+++ OloEngine/src/CMakeLists.txt
@@ -1042,6 +1042,7 @@ if(WIN32)
 		"Platform/Windows/WindowsEvent.h"
 		"Platform/Windows/WindowsSemaphore.h"
 		"Platform/Windows/WindowsInput.cpp"
+		"Platform/Windows/WindowsPlatformProcess.cpp"
 		"Platform/Windows/WindowsWindow.cpp"
 		"Platform/Windows/WindowsWindow.h"
 		"Platform/Windows/WindowsPlatformUtils.cpp"
@@ -1051,6 +1052,7 @@ elseif(UNIX AND NOT APPLE)
 		"Platform/Linux/LinuxEvent.h"
 		"Platform/Linux/LinuxSemaphore.h"
 		"Platform/Linux/LinuxInput.cpp"
+		"Platform/Linux/LinuxPlatformProcess.cpp"
 		"Platform/Linux/LinuxWindow.cpp"
 		"Platform/Linux/LinuxWindow.h"
 		"Platform/Linux/LinuxPlatformUtils.cpp"
